+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `nostalgie/internal/server` to `nostalgie/internal/renderer` to better reflect purpose.
 
 ## [0.68.3] - 2021-02-04
 ### Added